chore: cargo clippy --workspace is clean
`cargo clippy --workspace --all-targets` went from 826 warnings and 4 errors to
none; so does the Linux production build, `cargo clippy -p neomacs
--features video`. `cargo check` passes for the `webview`, allocator,
`--no-default-features`, and neovm-core `sound`/`vm-profile`/`fuzzing`/
`display-stop-validate` feature sets.

Most of it is mechanical: `assert_eq!(x, true)`, needless borrows, identity
conversions and casts, collapsible `if`s, doc-list indentation, and struct
literals in place of field reassignment after `Default::default()`. Code that
nothing reaches in any build was removed rather than silenced: an uncalled
backtrace-dispatch chain in `eval/apply.rs`, nine navigation builtins
superseded by other implementations, `SubrEntry::name_id` (nothing reads it,
and nothing reads `SubrEntry` by offset), and two stack slots every
JIT-compiled function allocated and never read.

Where the lint's own fix would cost something, the crate's existing
convention stands instead: four layout-engine enums keep
`#[allow(clippy::large_enum_variant)]`, since boxing them would add a heap
allocation per layout attempt; `BootstrapFont::Gui`, built once per startup,
is boxed. Two parity fixtures keep literal invisible characters under an
allow, because they live in raw strings where an escape is not processed.

An adversarial review of the whole diff (13 groups, each finding verified)
caught what a warning count cannot. Fixed here:

* A register restore in the regex engine was rewritten as
  `regstart[1..num_regs].copy_from_slice(..)`, which panics for a POSIX
  pattern without groups: its register vectors are empty and `num_regs` is 1.
  The index loop it replaced never touched them.
* Two `video`-only uses -- a `mut` binding and a test import -- looked unused
  under default features, so the production build stopped compiling.
* Two parity fixtures had invisible characters replaced by escape TEXT inside
  raw strings, changing the Lisp they send to GNU.
* Three parked powershell parity cases had been re-registered, and two
  `clone_and_copy` tests had stopped calling `Clone`.

Gate: the whole neovm-core suite passes in both JIT configurations,
10146/10146 each. The layout-engine and xtask suites have 8 failures that
predate this: their test bodies and the ci.yml they assert on are identical
to the pre-cleanup tree, and the layout two resolve fonts from whatever the
host has installed.

Measured (instructions, CPU 2), since this removed dead code beside hot paths
and two slots from every JIT frame: elisp-benchmarks moves within ±0.9% and
mostly downward -- elb-bytecomp -0.85%, elb-scroll -0.88%, map-closure
-0.29%, flet -0.05% (repeat -0.22%), elb-pcase +0.40% the only row above
+0.2%. Editor rows are flat: org-editing-heavy -0.08%, org-editing -0.00%,
indentation -0.17%.

+++ crates/neovm-core/src/emacs_core/display/xdisp/tests/mod.rs
@@ -1944,7 +1944,7 @@ fn window_text_pixel_size_counts_tty_wrapped_screen_rows() {
         .get_mut(buf_id)
         .expect("buffer")
         .insert("123456789012\nnext\n");
-    let selected_window = eval.frames.get(frame_id).expect("frame").selected_window.0 as u64;
+    let selected_window = eval.frames.get(frame_id).expect("frame").selected_window.0;
 
     let result = builtin_window_text_pixel_size_ctx(
         &mut eval,
@@ -1988,7 +1988,7 @@ fn window_text_pixel_size_honors_tty_truncate_lines() {
         .get_mut(buf_id)
         .expect("buffer")
         .insert("123456789012345678901234567890\nnext");
-    let selected_window = eval.frames.get(frame_id).expect("frame").selected_window.0 as u64;
+    let selected_window = eval.frames.get(frame_id).expect("frame").selected_window.0;
 
     let result = builtin_window_text_pixel_size_ctx(
         &mut eval,
@@ -5270,12 +5270,12 @@ fn test_vertical_motion_goal_column_past_row_end_lands_on_the_row_end_like_gnu()
             .map(|index| crate::window::DisplayPointSnapshot {
                 role: crate::window::DisplayPointRole::Glyph,
                 buffer_pos: crate::buffer::LispCharPos1::new(1 + index),
-                x: 8 * index as i64,
+                x: 8 * index,
                 y: 0,
                 width: 8,
                 height: 16,
                 row: 0,
-                col: index as i64,
+                col: index,
             })
             .collect::<Vec<_>>();
         let frame = eval.frame_manager_mut().get_mut(frame_id).expect("frame");
